pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "easi"
version = "0.1.0"
description = "EASI - Embodied Reasoning Evaluation for Spatial Intelligence"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "datasets>=2.0.0",
    "huggingface_hub>=0.19.0",
    "litellm>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
]
vllm = [
    "vllm>=0.15.1",
    "transformers>=4.40.0,<5.0"
]
custom-models = [
    "torch==2.7.1",
    "torchvision==0.22.1",
    "transformers>=4.40.0,<5.0",
    "accelerate>=0.26.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "Pillow>=10.0",
]

[project.scripts]
easi = "easi.cli:main"

[tool.setuptools.packages.find]
include = ["easi*"]
exclude = ["VLMEvalKit*", "lmms-eval*"]

[tool.setuptools.package-data]
"easi" = [
    "simulators/*/manifest.yaml",
    "simulators/*/*/conda_env.yaml",
    "simulators/*/*/requirements.txt",
    "tasks/*/*.yaml",
    "llm/models/*/manifest.yaml",
]
